requirements.txt
azureml-defaults==1.53.0
inference-schema[numpy-support]==1.7.0
xlrd==2.0.1
mlflow== 2.4.1
azureml-mlflow==1.51.0
scikit-learn
scipy
pandas
keras
seaborn
transformers==4.18.0
nltk
torch
xformers
azure.ai.ml
azure.identity
xgboost